requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langsmith>=0.1.0

# Vector stores and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
ollama>=0.1.0
langchain-ollama>=0.1.0
requests>=2.31.0

# Document processing
pypdf>=3.0.0
python-dotenv>=1.0.0

# Evaluation
datasets>=2.14.0
ragas>=0.1.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
rich>=13.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0

# Optional but recommended
jupyter>=1.0.0
ipykernel>=6.25.0
nest-asyncio>=1.5.0

# Web search (optional)
tavily-python>=0.3.0
langchain-tavily>=0.1.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sse-starlette>=1.8.2

